+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 * Don't color face when there is text property (629a21184552687f5b2ea7a378f0c363f0bd379a)
 * Accept cons cell for `backends` variables (1f628103d24c8871396da1405915b5104c26a203)
 * Calculate string length in pixel instead (bf375ac2b659e0312bc5a3a67aded4d3569c0b51)
+* Add capability to display backend name (#3)
 
 ## 0.1.1
 > Released Jun 15, 2022